requirements.txt
flask==2.3.2
python-telegram-bot==20.3
requests==2.31.0
pandas==2.0.3
sqlalchemy==2.0.20
pyyaml==6.0.1